backend/requirements.txt
asgiref==3.8.1
Django==4.1.13
django-cors-headers==4.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.7.0
pycryptodome==3.21.0
PyJWT==2.9.0
pymongo==3.12.3
pytz==2024.2
sqlparse==0.2.4
tzdata==2024.2
